requirements.txt
flask
psycopg2
python-decouple
python-dotenv